% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calhypscore.R
\name{calhypscore}
\alias{calhypscore}
\title{Calculate hypoxia score}
\usage{
calhypscore(expr, geneset, class, index = 1)
}
\arguments{
\item{expr}{should be a data.frame or matrix}

\item{geneset}{a hypoxia gene set}

\item{class}{"continuous" or "count"}

\item{index}{from 1 and 10, and the default value is 1}
}
\value{
hypoxia score, high group, low group and median score
}
\description{
This function allows users to select different hypoxia gene sets to calculate the hypoxia score per sample, and the default parameter is RWnetHS.
}
\examples{
res<-calhypscore(ExpMatrix, Hypoxia.Sig, class='count', index = 1); res
}
